tv-chart: extract getVisibleRange; clean up derived values
